requirements.txt
aiohttp==3.8.4
aiomysql==0.0.20
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
astroid==2.3.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
Authlib==1.2.0
azure-core==1.26.4
azure-storage-blob==12.16.0
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
blis==0.7.10
boto3==1.28.30
botocore==1.31.30
Brotli==1.0.9
cached-property==1.5.1
cachetools==5.3.1
catalogue==2.0.9
celery==5.3.4
git+https://github.com/anar2706/celery-slack.git#egg=celery-slack
certifi==2019.11.28
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.27.2
coloredlogs==15.0.1
confection==0.1.1
ConfigArgParse==1.5.3
contourpy==1.1.0
cryptography==39.0.2
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
docutils==0.15.2
effdet==0.4.1
email-validator==1.3.1
emoji==2.8.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl
gevent==22.10.2
exceptiongroup==1.1.1
fastapi==0.94.1
fastapi-jwt-auth==0.5.0
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.3.3
fsspec==2023.6.0
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
grpcio==1.54.0
grpcio-tools==1.54.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.9.2
httpx==0.23.3
huggingface-hub==0.16.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
install==1.3.5
iopath==0.1.10
isodate==0.6.1
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonpath-ng==1.5.2
jsonpointer==1.10
jsonschema==3.2.0
langchain==0.0.209
lazy-object-proxy==1.4.3
lit==16.0.6
locust==2.15.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
maxminddb==2.2.0
mccabe==0.6.1
mdurl==0.1.2
memory-profiler==0.61.0
more-itertools==8.0.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
mysql-connector-python==8.0.26
nameparser==1.0.5
ndg-httpsclient==0.5.1
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==2.0.6
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
packaging==23.0
pandas==2.0.3
pathtools==0.1.2
pathy==0.10.2
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.2
peewee==3.15.4
Pillow==9.5.0
pluggy==1.0.0
ply==3.11
portalocker==2.7.0
preshed==3.0.8
protobuf==4.24.0
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.16.1
PyJWT==1.7.1
pylint==2.4.4
Pympler==1.0.1
PyMySQL==0.9.2
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pypdf==3.15.2
pypdfium2==4.18.0
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.15
python-multipart==0.0.5
python-pptx==0.6.21
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
redis==4.5.4
regex==2023.5.5
requests==2.28.2
rfc3986==1.5.0
rich==13.5.2
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.2
scipy==1.11.1
sentry-sdk==1.19.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
starlette==0.26.1
streamlit==1.25.0
sympy==1.12
text-unidecode==1.3
thinc==8.1.12
timm==0.9.5
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tiktoken==0.4.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.10.0
unstructured-inference==0.5.9
urllib3==1.26.15
uvicorn==0.21.1
validators==0.21.0
weaviate-client==3.22.1
Werkzeug==2.2.3
wrapt==1.11.2
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
zope.event==4.6
zope.interface==6.0
